data/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
openai>=1.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0
rich>=13.7.0
python-dotenv>=1.0.0
